% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtax.r
\docType{data}
\name{qtax}
\alias{qtax}
\title{Quarterly tax data.}
\format{Data frame with 1 row per state, per tax, per quarter.
\describe{
\item{stabbr}{state abbreviation}
\item{date}{first day of quarter}
\item{ic}{item code for the variable, as defined by Census Bureau}
\item{vname}{variable name (a memorable name for the item code, ic)}
\item{value}{amount in $ millions}
\item{vdesc}{variable description}
}}
\source{
Census Bureau API for 1994+ data, plus various Census Bureau historical files for earlier data.
}
\usage{
qtax
}
\description{
Quarterly tax revenue data for individual state governments and the U.S. in aggregate, by tax type, from 1963 (U.S.) through recent.
}
\examples{
  head(qtax)
  comment(qtax)
}
\keyword{datasets}
