% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtax.r
\docType{data}
\name{qtax}
\alias{qtax}
\title{qtax Quarterly tax data.}
\format{Data frame with 1 row per variable per date, NA records have been removed.
\describe{
\item{stabbr}{chr: state abbreviation}
\item{date}{date: first day of quarter}
\item{ic}{chr: item code for the variable, as defined by Census Bureau}
\item{vname}{chr: variable name}
\item{value}{double: $ amount, in units described in FOF documentation}
\item{vdesc}{chr: variable description}
}}
\source{
Census Bureau API for recent , plus various historial files.
}
\usage{
qtax
}
\description{
qtax Quarterly tax revenue data for state governments and the U.S. total, by tax type, from 1963 (U.S.) through recent.
}
\examples{
  head(qtax)
}
\keyword{datasets}
